refactor(raijin): extract initializer interfaces
