+++ yarn/raijin/src/index.ts
@@ -1,3 +1,3 @@
-export type { RunRaijinInitializerOptions } from './initializer.js'
+export type { RunRaijinInitializerOptions } from './initializer.interfaces.js'
 
 export { runRaijinInitializer } from './initializer.js'
