+++ yarn/raijin/src/runtime-download.interfaces.ts
@@ -1,0 +1,1 @@
+export type FetchLike = typeof fetch
