+++ yarn/raijin/src/runtime-installer.interfaces.ts
@@ -1,0 +1,6 @@
+import type { FetchLike } from './runtime-download.interfaces.js'
+
+export interface InstallRaijinRuntimeOptions {
+  cwd: string
+  fetchImpl: FetchLike
+}
