+++ yarn/raijin/src/runtime.interfaces.ts
@@ -1,0 +1,9 @@
+export interface RaijinRuntimeManifest {
+  assetName: string
+  assetUrl: string
+  packageName: string
+  schemaVersion: number
+  sha256: string
+  tagName: string
+  version: string
+}
