+++ yarn/raijin/src/yarn-command.interfaces.ts
@@ -1,0 +1,1 @@
+export type YarnCommandRunner = (args: Array<string>, cwd: string) => Promise<void>
